Schema and related fields updated

+++ EbaySharp/Entities/Sell/Fulfillment/Order/DeliveryCost.cs
@@ -2,8 +2,8 @@
 {
     public class DeliveryCost: OrderAmount
     {
-        public float? DiscountAmount { get; set; }
-        public float? HandlingCost { get; set; }
+        public double? DiscountAmount { get; set; }
+        public double? HandlingCost { get; set; }
         public OrderAmount ImportCharges { get; set; }
         public OrderAmount ShippingCost { get; set; }
         public OrderAmount ShippingIntermediationFee { get; set; }
